+++ packages/native/storage/lib/src/local/local_storage.android.dart
@@ -42,4 +42,8 @@ final class LocalStoragePlatformAndroid extends NativeLocalStoragePlatform {
     _storage.write(key.toJString(), value.toJString());
     return value;
   }
+
+  @override
+  List<String> get allKeys =>
+      _storage.getAllKeys().map((key) => key.toDartString()).toList();
 }